pyproject.toml
[project]
name = "quantist_api"
version = "1.1.0"
description = "Backend of Quantist.io"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "fastapi[standard]==0.115.5",
    "future==1.0.0",
    "google-cloud-bigquery==3.27.0",
    "jinja2==3.1.4",
    "kaleido==0.2.1",
    "orjson==3.10.12",
    "pandas[performance]==2.2.3",
    "plotly==5.24.1",
    "polars==1.15.0",
    "psycopg==3.2.3",
    "pyarrow==18.1.0",
    "python-dotenv==1.0.1",
    "scikit-learn==1.5.2",
    "sqlalchemy==2.0.36",
    "sqlalchemy-bigquery==1.12.0",
    "uvicorn[standard]==0.32.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "ruff==0.8.0",
]
